gateway/requirements.txt
# API Gateway Dependencies

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
bcrypt>=4.1.0
redis>=5.0.0
aiokafka>=0.10.0
httpx>=0.26.0
python-multipart>=0.0.6
